Enhance `db:*` tasks to run per-shard
This change reduces the number of patches required to make the built-in Rails `db` rake tasks work with shards.

Instead of iterating through each migration and running it on the appropriate shard, the task must be run once per shard. Each migration must now have `shard_tags` explictly set, which ensures that each migration is only run on shards that satisify the migrations shard tags.

This also changes some of the specs setup code, allow the entire schema to rebuilt between specs, when necessary, and before the entire suite.