Improve how db:create|drop and db:(create|drop)_all work
This makes the methods behaviour more like Rails, namely, when in development, create/drop test and dev. Everywhere else, just create/drop the databases for each shard for that environment.

This is important when running on beta or CI because we don't have permission to create/drop databases outside of the specific env config.

This also moves code into a `Cuddlefish::DatabaseTasks` object. It's still not tested but it feels worse now, because it's in a proper object. (It's just too late to write tests.)